+++ .vscode/launch.json
@@ -26,7 +26,7 @@
                 "--colors",
                 "--compilers",
                 "ts:ts-node/register,tsx:ts-node/register",
-                "${workspaceRoot}/test/server.test.ts"
+                "${workspaceRoot}/test/server/index.test.ts"
             ],
             "internalConsoleOptions": "openOnSessionStart"
         }
